PLN/requirements.txt
pandas
sentence-transformers
numpy
faiss-cpu
streamlit
PyPDF2